Create watcloud-website repo (#1675)
Create the (readonly) watcloud-website repo. A part of #996 

Also had to make a lot of infrastructure changes because there's a new
use case where we want to extract a secret from the secrets vault.

Also moved everything to self-hosted runners because we ran out of
GitHub hosted runner credits this month. It's unfortunate that they
don't have automatic fallback to self-hosted.