Improve website intro and compute cluster documentation (#1796)
This PR adds a lot of documentation improvements.

[Latest preview for this
PR](https://rgw-preview.watonomous.ca/WATonomous/infra-config/pr-1796/latest/website/)

- [ ] ~~Automatically generate an index for sub-pages. For example, the
`Compute Cluster` page should show a list of documents under it. This is
like
https://github.com/shuding/nextra/blob/2f001d897f2109b24600f66fe2009e5f788b980c/docs/pages/docs/docs-theme.mdx,
but auto-generated. We may be able to borrow this logic from our Blog
page.~~ We'll defer this to a future iteration. For now, we can just
have "Overview" pages under the category.