Install dev vm software on SLURM nodes (#2452)
This PR effectively renames `ubuntu_dev_vm_nodes` to `login_nodes` and
uses the original name for both SLURM and login nodes. In this way, the
SLURM environment can have the same software as the login nodes.

Note that for the autofs mounts (/mnt/wato-drive*), it appears that we
can't trigger a mount from within a SLURM job, only outside of it.