Replace mps with shard GRES in SLURM (#2585)
We discovered a limitation of the mps scheduler that it is only able to
work on a single GPU:

https://slurm.schedmd.com/gres.html#MPS_Management
> Job requests for MPS will be processed the same as any other GRES
except that the request must be satisfied using only one GPU per node
and only one GPU per node may be configured for use with MPS.

https://github.com/basvandervlies/surf_slurm_mps?tab=readme-ov-file
> SLURM has implemented MPS support with the restriction that it only
supports 1 GPU in the node. We have multiple GPUs per node and we want
to use all the GPUs in this node for sharing.

This is a custom SPANK plugin for scheduling MPS jobs.

https://groups.google.com/g/slurm-users/c/3axnt5Wshkg
> MPS only works for the first GPU in a system. If you have a server
with multiple GPUs, you can only share the first GPU between multiple
jobs.

[Sharding](https://slurm.schedmd.com/gres.html#Sharding) works for all
GPUs.

It seems that shard is a nonstrict form of mps that works on all GPUs.

This PR switches scheduling to `shard` so that it works on all GPUs. The
only downside is that we don't get the benefits of MPS that helps
restrict processing power per process.