Introduce `compute_dense` SLURM partition for long-running jobs (#2816)
This PR introduces a `compute_dense` SLURM partition to handle jobs that
are long-running. This partition is set to have a maximum time limit of
7 days.

This PR also reverts WATonomous/infra-config#2815

1. The 6 hour limit is currently just a warning. It currently has no
effect on job scheduling. It was put in place to encourage workloads
that don't unnecessarily take up cluster resources if people forget to
shut them off. The limit wasn't enforced to account for people learning
to use SLURM any may want to run long-running jobs in interactive
shells.
2. 30-day limit is way too long as a general limit. The 1-day limit was
put in place to ensure there's no cluster lock-up while people get
familiar with SLURM and we learn about the usage patterns. It was also
helpful in case we need to perform emergency cluster maintenance.