Add documentation about port forwarding when using the provisioner container (#2916)
This PR adds some docs on how to port forward from the provisioner
container. This is useful for e.g. port forwarding from Kubernetes. This
may also be useful for developing the website--we may want to combine
the website development environment with the provisioner environment in
the future.

This PR also removes `network_mode: host`. The host network mode doesn't
appear to work in docker rootless (the service in `docker run -it
--network host node npx serve` isn't accessible from the host but the
service in `docker run -it -p 3000:3000 node npx serve` is). This is
replaced with instructions for port forwarding, as described above.

Also adds some general information about the provisioner container.