Create "services" page on the website (#3020)
We used to have a web page that describes everything in the onboarding
form. However, it was removed in
https://github.com/WATonomous/infra-config/pull/2259 due to the lack of
maintainers. This PR attempts to bring it back because we have been
running into [spam filter
issues](https://discord.com/channels/478659303167885314/571386709233893406/1273843070131114046)
with our onboarding emails that include a lot of links.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>