Fix blog post title image display (#3233)
## Description

For some reason, the blurred placeholder image that nextjs generates can
be of a different size than the original image. The title image we are
using is `1792 × 1024 px` (aspect ratio 1.75) but the blurred image is
`240 x 150 px` (aspect ratio 1.6). Perhaps nextjs uses a few fixed
sizes. When we switched from `cover` to `contain` for the blurred image
in #3232, it resulted in a noticeable flicker. This PR takes a different
approach to fixing the layout problem. This fixes both the layout shift
and makes sure the image clickbox is correct at all times.

Before/after:


https://github.com/user-attachments/assets/8ea2732c-285d-4e48-9489-b8687f436288


## Checklist
- [x] I have read and understood the [WATcloud
Guidelines](https://cloud.watonomous.ca/docs/community-docs/watcloud/guidelines)
- [x] I have performed a self-review of my code