Validate external links (#3441)
## Description

Introduces a script that checks whether page contains broken external
links, which are links that do not lead to intended external sites. This
includes pages that are hidden behind authentication. Links that might
be considered broken but that should be ignored can be hardcoded into
the script. A similar script was previously written to validate internal
links which resolved issue #2533.

Resolves #3441.
---------

Co-authored-by: Alex Boden <alex.boden@uwaterloo.ca>
Co-authored-by: Ben Zhang <ben-z@users.noreply.github.com>