Add `.generated` to `user.schema.json` (#3512)
## Description

We've had incidents where new members would edit `user.schema.json`
instead of `user.schema.jsonnet`. Initially we were hoping that
`user.schema.json` being inside `outputs/` was sufficient to signal that
this was an output, not a file to be edited manually, but it appears
that it's not clear enough. This PR renames it to
`user.schema.generated.json` so that it's more clear.


## Checklist
- [x] I have read and understood the [WATcloud
Guidelines](https://cloud.watonomous.ca/docs/community-docs/watcloud/guidelines)
- [x] I have performed a self-review of my code