Set onboarding form defaults only when necessary (#3707)
## Description

Currently, when a user fills out the onboarding form, new fields that
have default values are injected into the form data. This creates a lot
of verbosity. This PR makes it so that default values are not injected
unless they are required.

## Checklist
- [x] I have read and understood the [WATcloud
Guidelines](https://cloud.watonomous.ca/docs/community-docs/watcloud/guidelines)
- [x] I have performed a self-review of my code