Add watcloud-website to whitelist for external links (#4064)
## Description

<!--- Please provide a summary of your changes. Make sure to include
relevant motivation, context, and link related documents/conversations.
-->

<!--- Use [linking
keywords](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue)
like `Resolves #<issue_number>` to automatically close issues. -->

When creating a new page in the website the edit link is not in the
master repo so it gives a 404 error during the check_external_links job:
https://github.com/WATonomous/infra-config/actions/runs/14293366346/job/40057655930

This is link is to https://github.com/WATonomous/watcloud-website , we
can whitelist this.

## Checklist
- [x] I have read and understood the [WATcloud
Guidelines](https://cloud.watonomous.ca/docs/community-docs/watcloud/guidelines)
- [x] I have performed a self-review of my code